Create README.md
I definitely had an off day today. I was able to create the shapes but i'm still a little confused with the concept of function and definitely need more practice with them. Unfortunately, I wasn't able to get an of the exercises done. Overall, I wrote out my functions but I wasn't able to get them to work before the lab ended.